Agregando lógica de creación de salas, tipoSalas y asientos (Falta solucionar problema de ciclo que causa la dependencia en las eliminaciones).

+++ src/main/java/hn/unah/lenguajes1900/data/backend_proyecto_lenguajes_cine/services/impls/AsientoServiceImpl.java
@@ -50,8 +50,13 @@ public class AsientoServiceImpl implements AsientoService{
     public String eliminarAsientos(long codigoSala) {
 
         if(this.salaRepository.existsById(codigoSala)){
-            for (Asiento asiento : findAllByCodigoSala(codigoSala)) {
-                this.asientoRepository.delete(asiento);
+            List<Asiento> asientos = (List<Asiento>) this.asientoRepository.findAll();
+
+            for (Asiento asiento : asientos) {
+                if(asiento.getSala().getCodigoSala() == codigoSala){
+
+                    this.asientoRepository.delete(asiento);
+                }
             }
             return "Asientos eliminados.";
         }
